Automatically initialize the random number seed when needed.
A fair number of places relies on the random number generator build into LR
using the rand() function. That random number generator is not seeded with
something sufficiently random - the end user is supposed to do that.

This change will cause the random number generator to automatically get seeded
when needed when one of the functions that needs random numbers gets called.
There is a new function that is used to accomplish this: y_rand().

In order to make room for that function some renames happened:
y_rand_between() -> y_rand_in_sliding_window()
y_rand -> y_rand_between()

Both of those were more or less misnamed. Users of these functions need to
update their call sites as appropriate.