Add VTS_report_error(). Generic error reporting facility.
This probably fixes a few pretty bad bugs introduced in some of the earlier
patches in this series. (Look at the removed code in VTS_process_returncode
for instance..)

Now all I need to do is rip out all of the custom error handling by
replacing it with calls to VTS_report_error() or VTS_process_returncode().

Some additional cleanups and documentation changes included - for FREE!!!
with this patch.