Fix off by one error when changing the action prefix.
Starting a new action block will set the action prefix - a bit of text that is prefixed to the transaction names. When the action prefix was being set one character too few was copied, causing the null byte at the end of the string to not get copied to the correct position. By pure chance this code still worked for the first action block started or for cases where each subsequent action block had a prefix that was least as long as the previous one - the memory is prefilled with zeroes when allocated. However, as soon as a short prefix is copied on top of a longer one the missing null byte meant that only the start of the prefix was overwritten with the new prefix text. All characters beyond that point stayed where they were.

Tsk.