Make y_start_action_block() stop calling y_end_action_block() and stop y_end_action_block() from checking if there is a block transaction.
This makes the the optional block transaction code stop calling y_end_action_block() if the action prefix is still set. The drawback is that we no longer guarantee that y_end_action_block() is always run before the next action block starts, but this way having it set up a transaction for the block is more reliable.