Added y_calculate_thinktime_for_rampup()
For simulating situations with limited amounts of connections on the client side. 

In such a case we cannot use regular vuser based rampups, so instead we have to gradually lower the thinktime to get a similar effect. 
This will linearly decrease the thinktime based on the current time, until rampup_period has passed, at which point the resulting value will be zero.