update y_vts.c to use LR 11.52 interfaces.
In fact, this removes most of the y_vts.c code.

This library mostly added connection management for the situation where a script only has one VTS connection.
With the new LR 11.52 interface loadrunner gained a whole set of methods that did the exact same thing. As a result, all of the duplicate methods in y_vts.c have been removed.

I realize that this will break support for older loadrunner versions, but seriously, this code never was that good to begin with. Admitting it was crap and removing most of it is the right thing to do.